move dht record code into new package
